Migrate to AGP 8.1.0, Gradle 8.2.1, Kotlin 1.9.0, and KSP
