+++ app/src/main/kotlin/com/skydoves/pokedex/ui/details/DetailViewModel.kt
@@ -23,8 +23,8 @@ import androidx.lifecycle.viewModelScope
 import com.skydoves.bindables.BindingViewModel
 import com.skydoves.bindables.asBindingProperty
 import com.skydoves.bindables.bindingProperty
-import com.skydoves.core.data.repository.DetailRepository
 import com.skydoves.pokedex.core.model.PokemonInfo
+import com.skydoves.pokedex.core.repository.DetailRepository
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 import kotlinx.coroutines.flow.Flow
