+++ app/src/main/kotlin/com/skydoves/pokedex/ui/main/MainViewModel.kt
@@ -22,8 +22,8 @@ import androidx.lifecycle.viewModelScope
 import com.skydoves.bindables.BindingViewModel
 import com.skydoves.bindables.asBindingProperty
 import com.skydoves.bindables.bindingProperty
-import com.skydoves.core.data.repository.MainRepository
 import com.skydoves.pokedex.core.model.Pokemon
+import com.skydoves.pokedex.core.repository.MainRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.flatMapLatest
