+++ app/src/test/kotlin/com/skydoves/pokedex/MainViewModelTest.kt
@@ -21,12 +21,12 @@ import com.nhaarman.mockitokotlin2.atLeastOnce
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
-import com.skydoves.core.data.repository.MainRepository
-import com.skydoves.core.data.repository.MainRepositoryImpl
 import com.skydoves.pokedex.core.database.PokemonDao
 import com.skydoves.pokedex.core.database.entitiy.mapper.asEntity
 import com.skydoves.pokedex.core.network.service.PokedexClient
 import com.skydoves.pokedex.core.network.service.PokedexService
+import com.skydoves.pokedex.core.repository.MainRepository
+import com.skydoves.pokedex.core.repository.MainRepositoryImpl
 import com.skydoves.pokedex.core.test.MainCoroutinesRule
 import com.skydoves.pokedex.core.test.MockUtil
 import com.skydoves.pokedex.ui.main.MainViewModel
