+++ core-data/src/test/kotlin/com/skydoves/pokedex/core/data/DetailRepositoryTest.kt
@@ -22,11 +22,11 @@ import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
 import com.nhaarman.mockitokotlin2.whenever
-import com.skydoves.core.data.repository.DetailRepositoryImpl
 import com.skydoves.pokedex.core.database.PokemonInfoDao
 import com.skydoves.pokedex.core.database.entitiy.mapper.asEntity
 import com.skydoves.pokedex.core.network.service.PokedexClient
 import com.skydoves.pokedex.core.network.service.PokedexService
+import com.skydoves.pokedex.core.repository.DetailRepositoryImpl
 import com.skydoves.pokedex.core.test.MainCoroutinesRule
 import com.skydoves.pokedex.core.test.MockUtil.mockPokemonInfo
 import com.skydoves.sandwich.ApiResponse
