+++ gradle/libs.versions.toml
@@ -1,40 +1,39 @@
 [versions]
-agp = "7.4.2"
-kotlin = "1.8.10"
-ksp = "1.8.10-1.0.9"
-javaRelease = "8"
-jvmTarget = "1.8"
-material = "1.8.0"
+agp = "8.1.0"
+kotlin = "1.9.0"
+ksp = "1.9.0-1.0.12"
+jvmTarget = "17"
+material = "1.9.0"
 androidxActivity = "1.7.0"
-androidxFragment = "1.5.6"
+androidxFragment = "1.6.1"
 androidxLifecycle = "2.6.1"
-androidxRoom = "2.5.0"
-androidxArchCore = "2.1.0"
+androidxRoom = "2.5.2"
+androidxArchCore = "2.2.0"
 androidXStartup = "1.1.1"
 bindables = "1.1.0"
-hilt = "2.45"
+hilt = "2.47"
 retrofit = "2.9.0"
-okHttp = "4.10.0"
-sandwich = "1.3.5"
-moshi = "1.14.0"
-coroutines = "1.6.4"
+okHttp = "4.11.0"
+sandwich = "1.3.7"
+moshi = "1.15.0"
+coroutines = "1.7.2"
 whatIf = "1.1.2"
 glide = "4.15.1"
 glidePalette = "2.1.2"
 bundler = "1.0.4"
 transformationLayout = "1.1.2"
-recyclerView = "1.2.1"
+recyclerView = "1.3.1"
 baseAdapter = "1.0.4"
 androidRibbon = "1.0.4"
 progressView = "1.1.3"
 rainbow = "1.0.3"
 timber = "5.0.1"
-baselineProfiles = "1.3.0"
-macroBenchmark = "1.2.0-alpha12"
-uiAutomator = "2.3.0-alpha02"
-truth = "1.0.1"
+baselineProfiles = "1.3.1"
+macroBenchmark = "1.2.0-beta02"
+uiAutomator = "2.3.0-alpha04"
+truth = "1.1.3"
 junit = "4.13.1"
-turbine = "0.6.0"
+turbine = "0.13.0"
 robolectric = "4.4"
 androidxTest = "1.3.0"
 androidxTestJunit = "1.1.2"
